+++ src/components/CardContents.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import * as styles from "../styles/card.css";
+
+const CardContents = () => {
+  return (
+    <div className={styles.grid}>
+      <a className={styles.card} href="https://nextjs.org/docs" >
+        <h2>Documentation &rarr;</h2>
+        <p>Find in-depth information about Next.js features and API.</p>
+      </a>
+
+      <a className={styles.card} href="https://nextjs.org/learn" >
+        <h2>Learn &rarr;</h2>
+        <p>Learn about Next.js in an interactive course with quizzes!</p>
+      </a>
+
+      <a
+        href="https://github.com/vercel/next.js/tree/master/examples"
+        className={styles.card}
+      >
+        <h2>Examples &rarr;</h2>
+        <p>Discover and deploy boilerplate example Next.js projects.</p>
+      </a>
+
+      <a
+        href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+      >
+        <h2>Deploy &rarr;</h2>
+        <p>
+          Instantly deploy your Next.js site to a public URL with Vercel.
+        </p>
+      </a>
+    </div>
+  );
+};
+
+export default CardContents;
