+++ src/components/Footer.tsx
@@ -1,0 +1,22 @@
+import type {FC} from 'react';
+import * as styles from "../styles/footer.css";
+import Image from "next/image";
+
+const Footer:FC = () => {
+  return (
+    <footer className={styles.footer} >
+      <a
+        href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        Powered by{' '}
+        <span className={styles.logo}>
+            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
+          </span>
+      </a>
+    </footer>
+  );
+};
+
+export default Footer;
